refactor(server): extract agent runtime binding boundary
