+++ packages/server/src/__tests__/agent-service-extra.test.ts
@@ -13,7 +13,6 @@ import {
   clearAgentAvatarImage,
   createAgent,
   deleteAgent,
-  ensureClientSupportsRuntimeProvider,
   getAgentAvatarImage,
   getAgentSkills,
   listAgentsForAdmin,
@@ -25,6 +24,7 @@ import {
   updateAgent,
   updateAgentSkills,
 } from "../services/agent.js";
+import { ensureClientSupportsRuntimeProvider } from "../services/agent-runtime-binding.js";
 import { createMember } from "../services/member.js";
 import { createOrganization } from "../services/organization.js";
 import { createAdminContext, useTestApp } from "./helpers.js";
